New design to setup warehouse service with database cluster
