+++ pkgs/backend/src/entities/wms/stock_entries.rs
@@ -1,0 +1,22 @@
+//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.14
+
+use sea_orm::entity::prelude::*;
+
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
+#[sea_orm(table_name = "wms_stock_entries")]
+pub struct Model {
+    #[sea_orm(primary_key)]
+    pub id: i32,
+    pub tenant_id: Option<i32>,
+    pub lot_id: i32,
+    pub quantity: i32,
+    pub status: i32,
+    pub expired_at: Option<DateTimeUtc>,
+    pub created_at: Option<DateTimeUtc>,
+    pub updated_at: Option<DateTimeUtc>,
+}
+
+#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+pub enum Relation {}
+
+impl ActiveModelBehavior for ActiveModel {}
